Move a few more things around
