+++ CreeDictionary/API/search/types.py
@@ -1,9 +1,14 @@
-from typing import NamedTuple, Union, Set, NewType, Iterable, Tuple
+from typing import NamedTuple, Union, Set, NewType, Iterable, Tuple, cast
+
+import attr
+from attr import attrs
 
 from typing_extensions import Protocol
 
-from API.models import Wordform
-from API.schema import SerializedLinguisticTag
+from API.helpers import serialize_definitions
+from API.models import Wordform, Definition
+from API.schema import SerializedLinguisticTag, SerializedSearchResult
+from utils import Language
 from utils.fst_analysis_parser import LABELS
 from utils.types import ConcatAnalysis, FSTTag
 
@@ -131,3 +136,88 @@ def linguistic_tag_from_fst_tags(tags: Tuple[FSTTag, ...]) -> LinguisticTag:
         return SimpleLinguisticTag(tags[0])
     else:
         return CompoundLinguisticTag(tags)
+
+
+@attrs(auto_attribs=True, frozen=True)  # frozen makes it hashable
+class SearchResult:
+    """
+    Contains all of the information needed to display a search result.
+
+    Comment:
+    Each instance corresponds visually to one card shown on the interface
+    """
+
+    # the text of the match
+    matched_cree: str
+
+    is_lemma: bool
+
+    # English or Cree
+    matched_by: Language
+
+    # The matched lemma
+    lemma_wordform: Wordform
+
+    # triple dots in type annotation means they can be empty
+
+    # user friendly linguistic breakdowns
+    linguistic_breakdown_head: Tuple[str, ...]
+    linguistic_breakdown_tail: Tuple[str, ...]
+
+    # The suffix tags, straight from the FST
+    raw_suffix_tags: Tuple[FSTTag, ...]
+
+    # Sequence of all preverb tags, in order
+    # Optional: we might not have some preverbs in our database
+    preverbs: Tuple[Preverb, ...]
+
+    # TODO: there are things to be figured out for this :/
+    # Sequence of all reduplication tags present, in order
+    reduplication_tags: Tuple[str, ...]
+    # Sequence of all initial change tags
+    initial_change_tags: Tuple[str, ...]
+
+    definitions: Tuple[Definition, ...]
+
+    def serialize(self, include_auto_definitions) -> SerializedSearchResult:
+        """
+        serialize the instance, can be used before passing into a template / in an API view, etc.
+        """
+        # note: passing in serialized "dumb" object instead of smart ones to templates is a Django best practice
+        # it avoids unnecessary database access and makes APIs easier to create
+        result = attr.asdict(self)
+        # lemma field will refer to lemma_wordform itself, which makes it impossible to serialize
+        result["lemma_wordform"] = self.lemma_wordform.serialize(
+            include_auto_definitions=include_auto_definitions
+        )
+
+        result["preverbs"] = []
+        for pv in self.preverbs:
+            if isinstance(pv, str):
+                result["preverbs"].append(pv)
+            else:  # Wordform
+                result["preverbs"].append(
+                    pv.serialize(include_auto_definitions=include_auto_definitions)
+                )
+
+        result["matched_by"] = self.matched_by.name
+        result["definitions"] = serialize_definitions(
+            self.definitions, include_auto_definitions
+        )
+        result["relevant_tags"] = tuple(t.serialize() for t in self.relevant_tags)
+
+        return cast(SerializedSearchResult, result)
+
+    @property
+    def relevant_tags(self) -> Tuple[LinguisticTag, ...]:
+        """
+        Tags and features to display in the linguistic breakdown pop-up.
+        This omits preverbs and other features displayed elsewhere
+
+        In itwêwina, these tags are derived from the suffix features exclusively.
+        We chunk based on the English relabelleings!
+        """
+        return tuple(
+            linguistic_tag_from_fst_tags(fst_tags)
+            for fst_tags in LABELS.english.chunk(self.raw_suffix_tags)
+        )
